Changed T to lower case in Write the Docs
OK, in fairness, this wasn’t my fault. The title of the article I was
linking to uses a capitalized T. So because we have a stupid robot
overlord that doesn’t allow me to actually quote things correctly, I
was faced with two terrible options: Add “(sic)” to the title or the
article or misquote the article. I chose the latter, though, really,
the bot should be altered.