Make the templates for 2018 generic.
This was done when adding the Cincy conf to the site. Now we don’t have city-specific templating, but just a base set of templates for each year, except for the menus (for now).

This allows us to generate all the interesting bits & pieces from the YAML file, without having to customize each set of templates.