Add a Rules section to Slack (#840)
* Add a Rules section to Slack

This adds rules that are more explicit than our social rules.
It also addresses the Pitching & Brand issues we've had recently.

* edit mostly for content organization, IA

Changed order of things because I think the social rules and black sheep take precedence. We need the no sales-y content too, but the other stuff seems to me to affect more users. I also thought that putting the inclusive language bit in with the no sales-y content bits diluted the message.
We have more work to do here, but this commit represents my first concerns about the original PR; I agree with @mxsasha and @mjang that we need to be clearer about the CoC also but this is as much as I can manage atm. (With apologies for any typos etc; I don't usually use the GH UI ...)

* Language tweaks. Slightly more positive, bit less informal, fewer idioms.

* More explicit reference to CoC early in the text

* Add note about animations

* Update docs/slack.rst