Suggestion: update MMS -> Writing Style Guide (#1263)
https://docs.microsoft.com/en-us/style-guide/welcome/

From the guide:
"The Microsoft Writing Style Guide replaces the Microsoft Manual of Style, a respected source of editorial guidance for the tech community for more than 20 years. The style guide features updated direction and new guidance for subjects that weren't around when the last edition released. But it's also a reimagining of Microsoft style—a tool to help everyone write in a way that's natural, simple, and clear."