Add 2020 stats and community update
The community update got a bit long,
I"m wondering if we should split this out into two posts:

* 2020 Stats
* 2021 Kickoff

Where the 2021 kickoff goes to the newsletter,
and the stats are just linked.

We did this in 2018, but then did a combined update the last couple years.
I don't feel strongly either way.