Slack responders (#1424)
* Slack responders

* Replacing responder text with descriptions

Thanks for getting the ball rolling on this one! I took out the text of the actual replies and put in descriptions instead, so we have a little more wiggle room to edit the responses without having to remember to update the website too. (This is particularly important for the invite link, which expires periodically.)

Co-authored-by: Janine Chan <64388808+janine-c@users.noreply.github.com>