Update beginners-guide-to-docs.rst
beginners-guide-to-docs.rst is very specific to the developer audience but wasn't clearly labeled/branded that way. Also, it read like a presentation instead of feeling clear, actionable, and to the point. This commit updates the content so it is 100% obvious, at a glance, what the content is about and who the audience is.

Some areas of the content were edited to sound less like a manifesto. Also, the commit tries to suppress the presentation-style rhetoric to streamline for a written conversation that gets to the point faster.

Sentences were shortened for clarity and content was reorganized.