Update beginner's guide to docs (#2174)
* Update beginners-guide-to-docs.rst

beginners-guide-to-docs.rst is very specific to the developer audience but wasn't clearly labeled/branded that way. Also, it read like a presentation instead of feeling clear, actionable, and to the point. This commit updates the content so it is 100% obvious, at a glance, what the content is about and who the audience is.

Some areas of the content were edited to sound less like a manifesto. Also, the commit tries to suppress the presentation-style rhetoric to streamline for a written conversation that gets to the point faster.

Sentences were shortened for clarity and content was reorganized.

* Update beginners-guide-to-docs.rst

* update headers from "code" to "software"

* Update beginners-guide-to-docs.rst

Remove the FAQ section and mention the drawbacks of FAQs in a different, related section.

* clarify the audience

Clarify the audience and area of focus in the first sentence of the doc.

* Update docs/guide/writing/beginners-guide-to-docs.rst

Co-authored-by: Sam <sam@writethedocs.org>

* Update docs/guide/writing/beginners-guide-to-docs.rst

Co-authored-by: Sam <sam@writethedocs.org>

* Update beginners-guide-to-docs.rst

---------

Co-authored-by: Sam <sam@writethedocs.org>