fix:  Fix links to editors in `guide/contributing.rst`. (#2217)
Links to `Sublime Text` and `Visual Studio Code` editors
were not being displayed as intended because the link
syntax was not being parsed by `sphinx` inside the boldfaced
text block.  Fix was to isolate links outside of boldfaced block.
This was tested in an alternate `sphinx` build environment
to confirm links now show up on page as intended.

An issue was not created about this.